backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
pillow==10.1.0

